Make uploading work again

+++ app/uploaded_files/models.py
@@ -9,8 +9,8 @@ class UploadedFile(models.Model):
         "upload_jobs.UploadJob", on_delete=models.CASCADE, related_name="uploaded_files"
     )
     filename = models.CharField(max_length=255)
-    size = models.IntegerField(default=0)
-    transferred = models.IntegerField(default=0)
+    size = models.BigIntegerField(default=0)
+    transferred = models.BigIntegerField(default=0)
     media_type = models.CharField(max_length=255, blank=True, null=False)
     checksum_server = models.CharField(max_length=255, blank=True, null=False)
     checksum_client = models.CharField(max_length=255, blank=True, null=False)
